moved context models to proper model files
