+++ corva/models/base.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+from corva.event.event import Event
+
+
+class BaseContext(BaseModel):
+    class Config:
+        arbitrary_types_allowed = True
+
+    event: Event
