+++ corva/models/scheduled.py
@@ -1,0 +1,6 @@
+from corva.models.base import BaseContext
+from corva.state.redis_state import RedisState
+
+
+class ScheduledContext(BaseContext):
+    state: RedisState
