+++ corva/models/task.py
@@ -2,6 +2,8 @@ from typing import Any, Dict, Literal, Optional
 
 from pydantic import BaseModel
 
+from corva.models.base import BaseContext
+
 
 class TaskData(BaseModel):
     id: str
@@ -18,3 +20,8 @@ class TaskData(BaseModel):
 class UpdateTaskInfoData(BaseModel):
     fail_reason: Optional[str] = None
     payload: dict = {}
+
+
+class TaskContext(BaseContext):
+    task: TaskData
+    save_data: dict = {}
